refactor: created audio-player as seperate component
